Moved logging functionality to dedicated file
